pyproject.toml
[project]
name = "linebot-inspiration-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.7"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "line-bot-sdk>=3.21.0",
    "python-dotenv>=1.2.1",
    "openai>=2.14.0",
    "pyngrok>=7.5.0",
]
